+++ event-to-sentry.go
@@ -7,11 +7,9 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 	"io/ioutil"
 	"log"
-	"math/rand"
 	"net/http"
 	"net/url"
 	"os"
-	"github.com/google/uuid"
 	"github.com/joho/godotenv"
 	"strings"
 	"time"
@@ -239,29 +237,29 @@ func main() {
 		var timestamper Timestamper 
 		var bodyEncoder BodyEncoder
 		var envelopeEncoder EnvelopeEncoder
-		var headerKeys []string
-		fmt.Println(headerKeys)
+		// var headerKeys []string
 		var storeEndpoint string
 		var requestBody []byte
 
 		if (event.Kind == "error") {			
-			body, timestamper, bodyEncoder, headerKeys, storeEndpoint = decodeEvent(event)
+			
+			body, timestamper, bodyEncoder, storeEndpoint = decodeError(event)
 			body = eventId(body)
 			body = release(body)
 			body = user(body)
 			body = timestamper(body, event.Platform)
+			
 			undertake(body)
 			requestBody = bodyEncoder(body)
+
 		} else if (event.Kind == "transaction") {
 			
-			envelope, timestamper, envelopeEncoder, headerKeys, storeEndpoint = decodeEnvelope(event)
-
-			fmt.Printf(" %T %T %T %T\n", timestamper, envelopeEncoder, headerKeys, storeEndpoint)
+			envelope, timestamper, envelopeEncoder, storeEndpoint = decodeEnvelope(event)
+			// fmt.Printf(" %T %T %T %T\n", timestamper, envelopeEncoder, storeEndpoint)
 
 			// TODO transform the envelope Array, update traceId, release, user, timestamps
+			
 			// undertaker()			
-			// requestBody = []byte(envelope)
-
 			requestBody = envelopeEncoder(envelope)
 		}
 
@@ -294,75 +292,26 @@ func main() {
 }
 
 func buildRequest(requestBody []byte, eventHeaders map[string]string, storeEndpoint string) *http.Request {
+	if requestBody == nil {
+		log.Fatalln("buildRequest missing requestBody")
+	}
+	if eventHeaders == nil {
+		log.Fatalln("buildRequest missing eventHeaders")
+	}
+	if storeEndpoint == "" {
+		log.Fatalln("buildRequest missing storeEndpoint")
+	}
+
 	request, errNewRequest := http.NewRequest("POST", storeEndpoint, bytes.NewReader(requestBody)) // &buf
 	if errNewRequest != nil {
 		log.Fatalln(errNewRequest)
 	}
 
 	for key, value := range eventHeaders {
-		// request.Header.Set(key, value)
-		fmt.Println("HEADER", key, value)
-		if (key == "X-Sentry-Auth") {
-			fmt.Println("ignore")
-		} else {
+		if (key != "X-Sentry-Auth") {
 			request.Header.Set(key, value)
 		}
 	}
 	return request
 }
 
-// same eventId cannot be accepted twice by Sentry
-func eventId(body map[string]interface{}) map[string]interface{} {
-	if _, ok := body["event_id"]; !ok {
-		log.Print("no event_id on object from DB")
-	}
-	var uuid4 = strings.ReplaceAll(uuid.New().String(), "-", "")
-	body["event_id"] = uuid4
-	fmt.Println("> event_id updated", body["event_id"])
-	return body
-}
-
-// CalVer-lite
-func release(body map[string]interface{}) map[string]interface{} {
-	date := time.Now()
-	month := date.Month()
-	day := date.Day()
-	var week int
-	switch {
-	case day <= 7:
-		week = 1
-	case day >= 8 && day <= 14:
-		week = 2
-	case day >= 15 && day <= 21:
-		week = 3
-	case day >= 22:
-		week = 4
-	}
-	release := fmt.Sprint(int(month), ".", week)
-	body["release"] = release
-	fmt.Println("> release", body["release"])
-	return body
-}
-
-func user(body map[string]interface{}) map[string]interface{} {
-	if body["user"] == nil {
-		body["user"] = make(map[string]interface{})
-		user := body["user"].(map[string]interface{})
-		rand.Seed(time.Now().UnixNano())
-		alpha := strings.Split("abcdefghijklmnopqrstuvwxyz", "")[rand.Intn(9)]
-		var alphanumeric string
-		for i := 0; i < 3; i++ {
-			alphanumeric += strings.Split("abcdefghijklmnopqrstuvwxyz0123456789", "")[rand.Intn(35)]
-		}
-		user["email"] = fmt.Sprint(alpha, alphanumeric, "@yahoo.com")
-	}
-	return body
-}
-
-func undertake(body map[string]interface{}) {
-	if body["tags"] == nil {
-		body["tags"] = make(map[string]interface{})
-	}
-	tags := body["tags"].(map[string]interface{})
-	tags["undertaker"] = "h4ckweek"
-}
